restructured and fixed compilation errors
