+++ SPT25-20-0200A/SPT25-20-0200A.cpp
@@ -1,4 +1,4 @@
-#include "SPT25_20_0200A.h"
+#include "SPT25-20-0200A.h"
 
 const char* arr[2] = {"kPa (Raw)", "kPa"};
 t_datasetup datasetup = {2, arr};
@@ -10,6 +10,10 @@ SPT25_20_0200A::SPT25_20_0200A(uint8_t pin, arduino_t arduino) : Sensor(S_SPTD25
 errorlevel_t SPT25_20_0200A::init(){
     pinMode(pin, INPUT);
     return ERR_NONE;
+}
+
+double spt25_20_200a_map(const int x, const double in_min, const double in_max, const double out_min, const double out_max) {
+  return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
 
 // TODO Calibrate from 0 psi to 200 psi
@@ -24,8 +28,4 @@ errorlevel_t SPT25_20_0200A::read(t_datum* data, uint8_t numdata){
     data[1].data = (float)spt25_20_200a_map(val, 176.0, 1023.0, 0, 200.0);
     // TODO: other error conditions?
     return ERR_NONE;
-}
-
-double spt25_20_200a_map(const int x, const double in_min, const double in_max, const double out_min, const double out_max) {
-  return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
